URL pointing to the sample metadata file, which is smaller and for test, demonstration, and vignette purposes only

URL pointing to the sample metadata file, which is smaller and for test, demonstration, and vignette purposes only

SAMPLE_DATABASE_URL

Format

An object of class character of length 1.

Examples

get_metadata(remote_url = SAMPLE_DATABASE_URL)
#> # Source:   table</github/home/.cache/R/CuratedAtlasQueryR/metadata.0.2.3.parquet> [?? x 56]
#> # Database: DuckDB 0.7.1 [unknown@Linux 5.15.0-1035-azure:R 4.2.3/:memory:]
#>    cell_               sample_   cell_type cell_type_harmonised confidence_class
#>    <chr>               <chr>     <chr>     <chr>                           <dbl>
#>  1 TTATGCTAGGGTGTTG_12 039c558c… mature N… immune_unclassified                 5
#>  2 GCTTGAACATGGTCTA_12 039c558c… mature N… cd8 tem                             3
#>  3 GTCATTTGTGCACCAC_12 039c558c… mature N… immune_unclassified                 5
#>  4 AAGGAGCGTATTCTCT_12 039c558c… mature N… immune_unclassified                 5
#>  5 ATTGGACTCCGTACAA_12 039c558c… mature N… immune_unclassified                 5
#>  6 CTCGTCAAGTACGTTC_12 039c558c… mature N… immune_unclassified                 5
#>  7 CGTCTACTCTCTAAGG_11 07e64957… mature N… immune_unclassified                 5
#>  8 ACGATCAAGGTGGTTG_15 17640030… mature N… immune_unclassified                 5
#>  9 TCATACTTCCGCACGA_15 17640030… mature N… immune_unclassified                 5
#> 10 GAGACCCTCCCTTCCC_15 17640030… mature N… immune_unclassified                 5
#> # ℹ more rows
#> # ℹ 51 more variables: cell_annotation_azimuth_l2 <chr>,
#> #   cell_annotation_blueprint_singler <chr>,
#> #   cell_annotation_monaco_singler <chr>, sample_id_db <chr>,
#> #   `_sample_name` <chr>, assay <chr>, assay_ontology_term_id <chr>,
#> #   file_id_db <chr>, cell_type_ontology_term_id <chr>,
#> #   development_stage <chr>, development_stage_ontology_term_id <chr>, …