Given a data frame of HCA metadata, returns a Seurat object corresponding to the samples in that data frame

Given a data frame of HCA metadata, returns a Seurat object corresponding to the samples in that data frame

get_seurat(...)

Arguments

...

Arguments passed on to get_SingleCellExperiment

data

A data frame containing, at minimum, a sample_ column, which corresponds to a single cell sample ID. This can be obtained from the get_metadata() function.

assays

A character vector whose elements must be either "counts" and/or "cpm", representing the corresponding assay(s) you want to request. By default only the count assay is downloaded. If you are interested in comparing a limited amount of genes, the "cpm" assay is more appropriate.

repository

A character vector of length one. If provided, it should be an HTTP URL pointing to the location where the single cell data is stored.

cache_directory

An optional character vector of length one. If provided, it should indicate a local file path where any remotely accessed files should be copied.

features

An optional character vector of features (ie genes) to return the counts for. By default counts for all features will be returned.

Value

A Seurat object containing the same data as a call to get_SingleCellExperiment.

Examples

meta <- get_metadata() |> head(2)
seurat <- get_seurat(meta)
#> ℹ Realising metadata.
#> ℹ Synchronising files
#> ℹ Downloading 0 files, totalling 0 GB
#> ℹ Reading files.
#> ℹ Compiling Single Cell Experiment.