Returns unharmonised metadata for selected datasets.

Various metadata fields are not common between datasets, so it does not make sense for these to live in the main metadata table. This function is a utility that allows easy fetching of this data if necessary.

get_unharmonised_metadata(
  dataset_ids,
  remote_url =
    "https://object-store.rc.nectar.org.au/v1/AUTH_06d6e008e3e642da99d806ba3ea629c5/unharmonised_metadata",
  cache_directory = get_default_cache_dir()
)

Arguments

dataset_ids

A character vector, where each entry is a dataset ID obtained from the $file_id column of the table returned from get_metadata()

remote_url

Optional character vector of length 1. An HTTP URL pointing to the root URL under which all the unharmonised dataset files are located.

cache_directory

Optional character vector of length 1. A file path on your local system to a directory (not a file) that will be used to store the unharmonised metadata files.

Value

A named list, where each name is a dataset file ID, and each value is a "lazy data frame", ie a tbl.

Examples

dataset = "838ea006-2369-4e2c-b426-b2a744a2b02b"
harmonised_meta = get_metadata() |> dplyr::filter(file_id == dataset) |> dplyr::collect()
unharmonised_meta = get_unharmonised_metadata(dataset)
unharmonised_tbl = dplyr::collect(unharmonised_meta[[dataset]])
dplyr::left_join(harmonised_meta, unharmonised_tbl, by=c("file_id", "cell_"))
#> # A tibble: 168,860 × 77
#>    cell_ sample_ cell_…¹ cell_…² confi…³ cell_…⁴ cell_…⁵ cell_…⁶ sampl…⁷ _samp…⁸
#>    <chr> <chr>   <chr>   <chr>     <dbl> <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 GGAC… f63cb4… L2/3-6… neuron        1 NA      NA      NA      168593… H21.33…
#>  2 TCAC… 0d4d1f… L2/3-6… neuron        1 NA      NA      NA      f7d747… H21.33…
#>  3 TCAG… 3e5a3b… L2/3-6… neuron        1 NA      NA      NA      3417a9… H20.33…
#>  4 TCAG… 7010a3… L2/3-6… neuron        1 NA      NA      NA      246a59… H20.33…
#>  5 AGCC… 82bb9a… L2/3-6… neuron        1 NA      NA      NA      7a8f35… H21.33…
#>  6 CCTC… a233eb… L2/3-6… neuron        1 NA      NA      NA      188243… H18.30…
#>  7 CTCG… 27f104… L2/3-6… neuron        1 NA      NA      NA      a62943… H20.33…
#>  8 AGCT… 0190a2… L2/3-6… neuron        1 NA      NA      NA      c508a8… H20.33…
#>  9 CTCG… 95d846… L2/3-6… neuron        1 NA      NA      NA      29285d… H21.33…
#> 10 AGTG… b0e1c5… L2/3-6… neuron        1 NA      NA      NA      cd7823… H21.33…
#> # … with 168,850 more rows, 67 more variables: assay <chr>,
#> #   assay_ontology_term_id <chr>, file_id_db <chr>,
#> #   cell_type_ontology_term_id <chr>, development_stage <chr>,
#> #   development_stage_ontology_term_id <chr>, disease <chr>,
#> #   disease_ontology_term_id <chr>, ethnicity <chr>,
#> #   ethnicity_ontology_term_id <chr>, experiment___ <chr>, file_id <chr>,
#> #   is_primary_data_x <chr>, organism <chr>, organism_ontology_term_id <chr>, …