Returns unharmonised metadata for a metadata query

Various metadata fields are not common between datasets, so it does not make sense for these to live in the main metadata table. This function is a utility that allows easy fetching of this data if necessary.

get_unharmonised_metadata(metadata, ...)

Arguments

metadata

A lazy data frame obtained from get_metadata(), filtered down to some cells of interest

...

Arguments passed on to get_unharmonised_dataset

dataset_id

A character vector, where each entry is a dataset ID obtained from the $file_id column of the table returned from get_metadata()

cells

An optional character vector of cell IDs. If provided, only metadata for those cells will be returned.

conn

An optional DuckDB connection object. If provided, it will re-use the existing connection instead of opening a new one.

remote_url

Optional character vector of length 1. An HTTP URL pointing to the root URL under which all the unharmonised dataset files are located.

cache_directory

Optional character vector of length 1. A file path on your local system to a directory (not a file) that will be used to store the unharmonised metadata files.

Value

A tibble with two columns:

file_id: the same file_id as the main metadata table obtained from get_metadata()

unharmonised: a nested tibble, with one row per cell in the input metadata, containing unharmonised metadata

Examples

harmonised <- get_metadata() |> dplyr::filter(tissue == "kidney blood vessel")
unharmonised <- get_unharmonised_metadata(harmonised)